Solved problem 3.
We tried a few different approaches, at first it did not seem to workout. However, we made a few modifications until we found the most efficient solution.

First we tried to store all the multiples in a list in which we would later removed the nonprime numbers. This approach was taking too long to provide all the numbers.

Then we loop from the smallest to the largest and if we find a number that is a multiple of our target number,  we grab the other diviser that when multiplied with the number we just found is equivalant to the target number.
Then we check if that number is prime with the IsPrime function. If it is, we have found the largest prime factor of the target number. So we print it to the console and return to our home planet.