Solved problem 6.
I used a for loop to fing the sum of the squares of the first 100 natural numbers.

I used a second loop to add numbers from 1 to 100. Then I multiplied the total sum by itself.

Lastly, I substracted the sum of the squares from the square of the sum in order to get the difference.